Refactor scroll event
- 스크롤 이벤트 위치 조정
- 반응형 디자인을 위한 메인 페이지 height 조정
